Labs/requirements.txt
azure-ai-documentintelligence
azure-ai-formrecognizer
azure-core
numpy
python-dotenv
